refactor(auth): extract shared password validation rule
